requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via gcsfs
aiosignal==1.4.0
    # via aiohttp
asttokens==3.0.0
    # via stack-data
atpublic==5.1
    # via bigframes
attrs==25.3.0
    # via aiohttp
bigframes==2.21.0
    # via -r requirements.in
cachetools==5.5.2
    # via google-auth
certifi==2025.8.3
    # via
    #   pyogrio
    #   pyproj
    #   requests
charset-normalizer==3.4.3
    # via requests
cloudpickle==3.1.1
    # via bigframes
comm==0.2.3
    # via ipywidgets
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
db-dtypes==1.4.3
    # via
    #   bigframes
    #   google-cloud-bigquery
    #   pandas-gbq
decorator==5.2.1
    # via
    #   gcsfs
    #   ipython
executing==2.2.1
    # via stack-data
fonttools==4.60.0
    # via matplotlib
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.9.0
    # via
    #   bigframes
    #   gcsfs
gcsfs==2025.9.0
    # via bigframes
geopandas==1.1.1
    # via bigframes
google-api-core[grpc]==2.25.1
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-connection
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-functions
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   pandas-gbq
google-auth==2.23.4
    # via
    #   bigframes
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-bigquery
    #   google-cloud-bigquery-connection
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-functions
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==1.2.2
    # via
    #   gcsfs
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery[bqstorage,pandas]==3.38.0
    # via
    #   bigframes
    #   pandas-gbq
google-cloud-bigquery-connection==1.18.3
    # via bigframes
google-cloud-bigquery-storage==2.33.1
    # via
    #   bigframes
    #   google-cloud-bigquery
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-functions==1.20.4
    # via bigframes
google-cloud-resource-manager==1.14.2
    # via bigframes
google-cloud-storage==2.10.0
    # via
    #   bigframes
    #   gcsfs
google-crc32c==1.7.1
    # via google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.70.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via
    #   bigframes
    #   google-cloud-bigquery-connection
    #   google-cloud-functions
    #   google-cloud-resource-manager
grpcio==1.75.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.75.0
    # via google-api-core
humanize==4.13.0
    # via bigframes
idna==3.10
    # via
    #   requests
    #   yarl
ipython==9.5.0
    # via ipywidgets
ipython-pygments-lexers==1.1.1
    # via ipython
ipywidgets==8.1.7
    # via bigframes
jedi==0.19.2
    # via ipython
jupyterlab-widgets==3.0.15
    # via ipywidgets
kiwisolver==1.4.9
    # via matplotlib
markdown-it-py==4.0.0
    # via rich
matplotlib==3.10.6
    # via bigframes
matplotlib-inline==0.1.7
    # via ipython
mdurl==0.1.2
    # via markdown-it-py
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
numpy==2.3.3
    # via
    #   bigframes
    #   contourpy
    #   db-dtypes
    #   geopandas
    #   matplotlib
    #   pandas
    #   pandas-gbq
    #   pyogrio
    #   shapely
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via
    #   db-dtypes
    #   geopandas
    #   google-cloud-bigquery
    #   matplotlib
    #   pandas-gbq
    #   pyogrio
pandas==2.3.2
    # via
    #   bigframes
    #   db-dtypes
    #   geopandas
    #   google-cloud-bigquery
    #   pandas-gbq
pandas-gbq==0.29.2
    # via
    #   bigframes
    #   google-cloud-bigquery
parso==0.8.5
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.3.0
    # via matplotlib
prompt-toolkit==3.0.52
    # via ipython
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-bigquery-connection
    #   google-cloud-bigquery-storage
    #   google-cloud-functions
    #   google-cloud-resource-manager
protobuf==6.32.1
    # via
    #   google-api-core
    #   google-cloud-bigquery-connection
    #   google-cloud-bigquery-storage
    #   google-cloud-functions
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==21.0.0
    # via
    #   bigframes
    #   db-dtypes
    #   google-cloud-bigquery
    #   pandas-gbq
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydata-google-auth==1.9.1
    # via
    #   bigframes
    #   pandas-gbq
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   rich
pyogrio==0.11.1
    # via geopandas
pyparsing==3.2.5
    # via matplotlib
pyproj==3.7.2
    # via geopandas
python-dateutil==2.9.0.post0
    # via
    #   bigframes
    #   google-cloud-bigquery
    #   matplotlib
    #   pandas
python-dotenv==1.1.1
    # via -r requirements.in
pytz==2025.2
    # via
    #   bigframes
    #   pandas
requests==2.32.5
    # via
    #   bigframes
    #   gcsfs
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==13.9.4
    # via bigframes
rsa==4.9.1
    # via google-auth
shapely==2.1.1
    # via
    #   bigframes
    #   geopandas
six==1.17.0
    # via python-dateutil
sqlglot==27.16.3
    # via bigframes
stack-data==0.6.3
    # via ipython
tabulate==0.9.0
    # via bigframes
toolz==1.0.0
    # via bigframes
traitlets==5.14.3
    # via
    #   ipython
    #   ipywidgets
    #   matplotlib-inline
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   bigframes
    #   grpcio
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
wcwidth==0.2.13
    # via prompt-toolkit
widgetsnbextension==4.0.14
    # via ipywidgets
yarl==1.20.1
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
